Replace utils.utils grab-bag with concern-named modules
`ricecooker.utils.utils` was a stutter holding seven unrelated helpers.
Split it by concern and repoint every import site:

- paths.py     — extract_path_ext, make_dir_if_needed
- storage.py   — get_hash, copy_file_to_storage
- validators.py — is_valid_url, is_valid_uuid_string
- VideoURLFormatError moves into its sole consumer, youtube.py

Pure move-and-reimport; function bodies are unchanged. utils.py is
deleted with no compat shim (internal helper path, per issue #692).

Closes #692
