+++ ricecooker/utils/archive_assets.py
@@ -21,11 +21,11 @@ import shutil
 from collections import deque
 
 from ricecooker.config import LOGGER
+from ricecooker.utils.paths import extract_path_ext
 from ricecooker.utils.pipeline.exceptions import ExpectedFileException
 from ricecooker.utils.pipeline.exceptions import InvalidFileException
 from ricecooker.utils.references import DEFAULT_MAPPERS
 from ricecooker.utils.references import is_external_url
-from ricecooker.utils.utils import extract_path_ext
 
 
 class ArchiveProcessor:
