+++ ricecooker/utils/pipeline/extract_metadata.py
@@ -1,9 +1,9 @@
 from le_utils.constants import file_formats
 from le_utils.constants import format_presets
 
+from ricecooker.utils.paths import extract_path_ext
 from ricecooker.utils.pipeline.context import ContentNodeMetadata
 from ricecooker.utils.pipeline.context import FileMetadata
-from ricecooker.utils.utils import extract_path_ext
 from ricecooker.utils.videos import extract_duration_of_media
 from ricecooker.utils.videos import guess_video_preset_by_resolution
