+++ tests/utils/test_extensions.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ricecooker.utils.utils import extract_path_ext
+from ricecooker.utils.paths import extract_path_ext
 
 # Tests generated by Claude Sonnet 3.7
